+++ crates/stackable-versioned-macros/src/attrs/field.rs
@@ -1,6 +1,7 @@
 use darling::{util::SpannedValue, Error, FromField, FromMeta};
 use k8s_version::Version;
-use syn::{Field, Ident};
+use proc_macro2::Span;
+use syn::{Field, Ident, Path};
 
 use crate::{attrs::container::ContainerAttributes, consts::DEPRECATED_PREFIX};
 
@@ -40,6 +41,16 @@ pub(crate) struct FieldAttributes {
 #[derive(Clone, Debug, FromMeta)]
 pub(crate) struct AddedAttributes {
     pub(crate) since: SpannedValue<Version>,
+
+    #[darling(rename = "default", default = "default_default_fn")]
+    pub(crate) default_fn: SpannedValue<Path>,
+}
+
+fn default_default_fn() -> SpannedValue<Path> {
+    SpannedValue::new(
+        syn::parse_str("std::default::Default::default").unwrap(),
+        Span::call_site(),
+    )
 }
 
 #[derive(Clone, Debug, FromMeta)]
@@ -139,6 +150,7 @@ impl FieldAttributes {
 
         // First, validate that the added version is less than the deprecated
         // version.
+        // NOTE (@Techassi): Is this already covered by the code below?
         if let (Some(added_version), Some(deprecated_version)) = (added_version, deprecated_version)
         {
             if added_version >= deprecated_version {